refactor: Show only news for a given week
News is, well, news. The news for the day, month or the year can be
moved to other pages where filters can be used to select a specific
period. Also, one day is too short a period. A week is a better
interval for getting a summary of what was seen and done.