APILoader - add a timeout for socket connections
Just occasionally the SSL handshake to the eBird servers will hang and
the API call will sit there forever. Adding timeout will raise an error
if this happens. There's no retry but since the checklist is the last
object to get saved any error means that it will be reloaded/added the
next time the APILoader runs.