Add State and County models
eBird's subnational1 and subnational2 levels are used to describe
adminstrative levels in countries large and small. The names are
rather unwieldy so the models were given nicer names, Region
(subnational1) and District (subnational2). However, two levels is not
sufficient for all the areas of interest to birders. First, Area was
added for groups of Locations. For groups of states (subnational1), or
counties (subnational2), however, the names District and Region are
perfect. So, two models, State and County were added, and for the
existing databases, records in District were migrated to County and
records in Region were migrated to District.

Now we have a sparse set of models which provides a more flexible way
of handling collections of administrative areas. This means the models
can more accurately describe a country, for example, for Portugal the
hierarchy of models for a location are:

    Location -> [Area] -> County -> District -> Region -> Country

While for large countries like the United States of America, the
hierarchy would be:

    Location -> [Area] -> County -> State -> Region -> Country

Area is optional. It could be used for locations within a national
park for example, or an island.

A District is notionally smaller than a State, but they are somewhat
interchangeable. They fulfill more or less the same role in the
hierarchy. However since the models are sparse, the redundancy is not
important. It might even be useful in modelling countries with very
large states, e.g. Texas, or California where the ecology and the bird
change significantly:

    Location -> [Area] -> County -> District -> State -> Region -> Country