Checklists - replace search panel with single search field
The search panel allowed checklists to be searched on country,
district, county, site and observer. Since country, district, county
and site are intimately linked - you can't search on a county that is
not part of the selected district or country. That means the search
simplifies to location and observer. In the context of the
application, searching for checklists based on location and/or
observer, is probably too fine grained. That means the search can be
further simplified to location or observer, which can be handled with
a single search field.

The bootstrap5-autocomplete used on the News page works well for this,
so we can get rid of django-filter and django-autocomplete-light. The
latter is important as a server request is triggered after every key
stroke, whereas bootstrap5-autocomplete issues one request for
everything when the page is loaded, and then filters on the client
side, which is far more efficient.

Now the checklists page is much simpler, and cleaner.