Add an alias for Observer name
Tour guides sometimes (often?) include the name of their company and a
web address (sigh) as their name. Commercial speech is not a problem
(mostly), rather it's the length of the string since that generates
ugliness if the string is wrapped.

The byname field is used for an edited version of the name, which will
be displayed, falling back to the name field if it is not set.