Copy 'started' attribute on Checklist to Observation
If a checklist does not set a time, for example if the observation
was incidental, then eBird sets the 'started' attribute to midnight,
i.e. the start of the day. That makes it a better candidate for sorting
observations as records will appear earlier. When using date and time,
observations with time equal to None will always appear later.