Only use County, State, and Country for locations
Being able to add areas for groups of Locations and Rergions for
groups of counties or state is a good idea, but it brings with it two
problems. First, the setup is more complicated, but not overly so. The
showstopper is maintainance. The current optimisation of mirroring the
foreign keys on Location to Checklist and Observation generates large
updates if any of the information changes. That's somewhat manageable,
the main issue is that anybody can add any name for a location, that
means they don't necessarily use the hotspot. That means maintaining
areas is going to be a lot of work if the feature is to be really
useful, since every new location needs to be reviewed to see if it
falls within the boundaries of an area. Whether the review was done
automatically or manually it's still a lot of work which really is
unsustainable. Sigh.