Rename the project to ebirders-pt
The site was started to need country specific changes - specifically
renaming locations to shorten them. The moderators for Portugal have a
specific naming scheme for grouping nearby sites in Important Bird
Areas or national parks. As a result the location names can be quite
long and that means trimming them wherever possible to keep the layout
of the web pages tidy and uncluttered.