Checklists - split search into separate fields
The single search field was tidy, however it reduced the utility of
the checklists page. In particular, given the high percentage of
incidental observations or low species counts mean that you have to
step through several pages to find useful information. By returning to
separate fields, search can be made the primary reason for having the
page. With a rich set of filters the checklists page can also be the
target when displaying, for example, the complete set of Big Lists for
a week - reducing the number of pages and code needed to provide a
rich set of features.