Latest/Weekly/Monthly News pages - added a High Counts table
The table shows highest count for each species seen during the period,
where the count is greater than the second highest count in the
previous month. The reason for taking the second highest count is
because generally eBirders don't make determined efforts to count
absolutely every individual, so taking the second count eliminates
records where they do. That way a single count cannot "suppress" all
subsequent high counts which are not as high.

It will be interesting to see whether there is a more analytical
approach, particularly after looking at the distributions for counts.
However for now, this heuristic seems to work really well, as it
generates a good number of records but not too many.