Observation - change approved and reviewed to be simple boolean fields
Having null for the approved and reviewed fields gave enough states to
support a colour coding system for flagging an observation (red,
yellow, or green). However the logic is ambiguous to read, so the null
values were dropped. When an observation is not marked as approved a
single (red) flag is displayed.

This will form the basis for an editorial policy where observations
are initially marked as approved (and not reviewed). Records where
there is doubt over the identification can then be marked as not
approved. After a period of time (30 days or the end of the month,
etc), records marked as not approved, but not reviewed can then be
checked and a final decision made.