Observation - remove remaining eBird Basic Dataset fields from the model
The fields were retained as the data could have been scraped from the
checklist web page. However the extra traffic and load on the ebird
servers was not really worth the extra value of the data
obtained. Also the goals of the site are more aligned towards
generating reports with observations and numbers rather than
population dynamics.