Checklist - rename 'coverage' field back to 'area'
The field was previously renamed to 'coverage' with the idea that an
Area model would be added to represent groups of Locations. Even with
well-known hotspots there are still a large number of private
locations added, so the effort required to maintain Areas was greater
than the value of having them, particularly with the "new plan" of
allowing multiple locations to be selected when performing searches.