Checklists Filter - Add a form for validating filterset fields
In theory, since the filter form only has choice fields there should
not be any errors, other than getting the from/until dates mixed up.
However if the user edits the querystring, then the default filterset
behaviour of ignoring the field is not sufficient. Adding a form lets
specific errors to be reported, particularly if there is an error is
generating the querystring when linking from other pages. It also adds
a layer to check for attempts to pass parameters of malicious intent.