Replace FilterView with a FormView
django-filter's FilterView works well enough if you just want to use
the basic/minimal configuration. However once you add a form for
reporting validation errors then the amount code is not very different
from adding a Form and changing the view to a FormView. The Form +
FormView is a cleaner fit with Django, gives you more precise control
over how the filtering and validation is performed and is easier to
write tests for.

django-filter appears to have a widget 'race-condition' problem. The
generated form, on first rendering, loses the widget attributes for
the OrderingFilter. This is because the OrderingFilter only accepts a
class and not an instance for the widget, which the other filter
fields accept either. That means you can only set the widget
attributes once the form is instantiated - by which time it's "too
late" to set the attributes. The easiest way to fix this is setting
them in the get_context_data() method of the view where you can access
the form for the filters directly.

Another, more serious problem for this application, where there are
three FilterViews sharing a basic set of filters, is that you cannot
define mixins then use them in a FilterSet. The filter fields lose
the widgets defined in the FilterSet and instead the default widgets
are displayed instead. There are two solutions, duplicate the code
and tests across all the filters, or drop django-filter. And, here
we are.