Checklist Filter - move filtering and ordering into the form
Since the form is used to define the fields used to filter and order
the list of objects, it is obvious/natural that responsibility for
creating the queryset filters and fields for the order_by clause is
handled by the form also. The simplifies the view since it does not
need to know the details of the form.

NOTE: The possibility of moving generating the queryset into the form
was also looked at, but then you'd have to move the pagination as well,
since both are the responsibility of the MultipleObjectMixin. It is
doable, but the mixin inherits from the ContentMixin so really should
only be used in a view. You could split the responsibilities of the
MultipleObjectMixin but you'd end up with more code to maintain.