Filter Forms - choices only need to be set when a value is submitted
Since the country, state, county, location, observer and species fields
are populated with autocomplete, there is no need to define the complete
set of choices when the field is rendered. The selected only needs to be
define when a value is submitted so the field shows the name of the item.