Convert the filter mixin into forms
Since there are no duplicate fields in any of the mixins used to filter
checklist or observations, and the widget for reach field extracts the
entry in the submitted data using the name of the field the mixins can
be converted to forms and rendered as sections inside <form> tags.
There is no need to add a prefix (which would have to happen if there
were duplicate names) so that helps keep the query parameter names
nice and short.

This also helps simplify the view, since the forms are instantiated,
bound, and validated on every GET request.