Simplify the tests
Rely, at least initially, on a simple set of tests for all the views.
This has a number of advantages:

Simple - does the view get loaded without errors. Also, pytest will
raise an error on any unset template variables so there test is
actually more powerful than first appears.

Efficient - too much time is wasted deciding what to test, and how. It
is trivial to write tests that simply test Django. It takes quite a
bit of thinking to write tests for only the code you add. Less time
spent writing, refactoring tests is more time that can be spent on
code quality, particularly simplifying code to rely more on
configuration, than custom methods.

Locality of reference - everything is in one place to it's much easier
to see when a view a missing.