News pages - Added postgreSQL materialized views
Added materialized views for the Year List, Big Week and Big Month
tables to reduce the number of Observerations to be processed. In the
case of the Year List this reduced the number from 750,000 (at time of
writing) down to 30,000. The query time reduced from 8 seconds to 50ms.

For the latest, weekly and monthly species totals the reduction in the
number of observations was more modest - about half but this still means
the query speed was cut a similar amount.

The Latest News page loads in about one second, without caching. With
caching it loads reliably in 75ms.

The Weekly News page loads in about 1.5 seconds, without caching. With
caching it loads reliably in 100ms.

The Monthly News page loads in about 2.3 seconds, without caching. With
caching it loads reliably in 100ms.