Project - Break apart the filters app
It's cleaner / more logical to have a base app for the parent view and
form, then move the filter forms into apps for the thing that is being
filtered on, locations, observers, etc. rather then having everthing
sit in a generic filters app. That makes it easy to apply changes
locally to the apps that contain the views, rather than try to
accomodate them in a shared app.