Replace YearList with CountryList, StateList and CountyList
Replacing the materialized view for the year list with separate views
for county, state and county greatly speeds up the loading of the
Species page. The new views are used for simple filters, falling back
to a full search of the Observations list for more complex queries.
However it is worth noting that more complex queries are working on a
much smaller set of observations so the overall effect is that all
queries are now much faster.